Alterando todos os ids e fks para tipo Long

+++ Back-end/techbridge/src/main/kotlin/com/techbridge/techbridge/dto/RespostaAventureiroDTO.kt
@@ -1,7 +1,7 @@
 package com.techbridge.techbridge.dto
 
 data class RespostaAventureiroDTO(
-    val idResposta: Int,
+    val idResposta: Long,
     val usuarioId: Long,
     val perguntaId: Int,
     val alternativaEscolhida: Int
